Man kann jetzt unter http://127.0.0.1:8050/user/state
den State zu einem User und einer Question abholen.
Achtung: gesaved mit JWT

Beispiel json:

{"q" : { "questionID" : 11,
				"question" : "Testfrage 1?",
"answer": "<p>Testantwort1</p>",
"topic" : {
"topicID": 3,
"topicName": "Wissenfragen",
"topicDescription": "Alle Fragen dieser Topic drehen sich um Datenbanken"
}},
 "user" : {
	 "email" : "marileen.stamer@stud.fh-luebeck.de",
	 "password" : "123"
 }
}

+++ src/main/java/lernapp/service/BasicService.java
@@ -52,14 +52,14 @@ public class BasicService<T> {
         return entity;
     }
 
-    /**
-     * Generic Method for saving A LIST OF ENTITIES of any types of Entities
-     * The Entity will be persisted to the database (not merged)
-     *
-     * @param  entities the Entity-List
-     * @return entity
-     *
-     */
+        /**
+         * Generic Method for saving A LIST OF ENTITIES of any types of Entities
+         * The Entity will be persisted to the database (not merged)
+         *
+         * @param  entities the Entity-List
+         * @return entity
+         *
+         */
     public <T> List<T> save (List<T> entities) {
 
         EntityManager em = EMF.createEntityManager();
@@ -97,12 +97,12 @@ public class BasicService<T> {
 
 
     // Query an entity by id
-    public T queryById(Long id, Class<T> entityType) {
+    public T queryById(int id, Class<T> entityType) {
         return EMF.createEntityManager().find(entityType, id);
     }
 
     // Delete entity by id
-    public void deleteById(Long id) {
+    public void deleteById(int id) {
         EntityManager em = EMF.createEntityManager();
         em.getTransaction().begin();
         Question result = em.find(Question.class, id);
